+++ bott/physics_models.py
@@ -1,1 +1,4 @@
 # Put the forward methods (like f(tilt, thickness) = PACBED)
+
+def simulate_cbed(thickness, tilt_x, tilt_y, params_abTEM=None):
+    pass